+++ packages/perseus-editor/src/widgets/interactive-graph-editor/components/graph-type-selector.tsx
@@ -5,6 +5,7 @@ import * as React from "react";
 type GraphTypeSelectorProps = {
     graphType: string;
     onChange: (newGraphType: string) => void;
+    showNoneOption: boolean;
 };
 
 const GraphTypeSelector = (props: GraphTypeSelectorProps) => {
@@ -12,9 +13,10 @@ const GraphTypeSelector = (props: GraphTypeSelectorProps) => {
         <SingleSelect
             selectedValue={props.graphType}
             onChange={props.onChange}
-            placeholder="Select a graph type"
+            placeholder="Select an answer type"
             style={styles.singleSelectShort}
         >
+            {props.showNoneOption && <OptionItem value="none" label="None" />}
             <OptionItem value="linear" label="Linear function" />
             <OptionItem value="quadratic" label="Quadratic function" />
             <OptionItem value="sinusoid" label="Sinusoid function" />