+++ packages/perseus-editor/src/widgets/interactive-graph-editor/locked-figures/util.test.ts
@@ -8,6 +8,7 @@ describe("getDefaultFigureForType", () => {
             coord: [0, 0],
             color: "grayH",
             filled: true,
+            labels: [],
         });
     });
 
@@ -22,18 +23,21 @@ describe("getDefaultFigureForType", () => {
                     coord: [0, 0],
                     color: "grayH",
                     filled: true,
+                    labels: [],
                 },
                 {
                     type: "point",
                     coord: [2, 2],
                     color: "grayH",
                     filled: true,
+                    labels: [],
                 },
             ],
             color: "grayH",
             lineStyle: "solid",
             showPoint1: false,
             showPoint2: false,
+            labels: [],
         });
     });