+++ packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords/start-coords-sinusoid.tsx
@@ -13,11 +13,13 @@ import CoordinatePairInput from "../../../components/coordinate-pair-input";
 
 import {getSinusoidEquation} from "./util";
 
-import type {Coord, PerseusGraphType} from "@khanacademy/perseus";
+import type {Coord} from "@khanacademy/perseus";
+
+type SinusoidCoords = [Coord, Coord];
 
 type Props = {
-    startCoords: [Coord, Coord];
-    onChange: (startCoords: PerseusGraphType["startCoords"]) => void;
+    startCoords: SinusoidCoords;
+    onChange: (startCoords: SinusoidCoords) => void;
 };
 
 const StartCoordsSinusoid = (props: Props) => {