+++ packages/perseus/src/widgets/expression/expression.types.ts
@@ -1,0 +1,9 @@
+import type {PerseusExpressionWidgetOptions} from "../../perseus-types";
+
+export type Rubric = PerseusExpressionWidgetOptions;
+
+export type OnInputErrorFunctionType = (
+    arg1?: any,
+    arg2?: any,
+    arg3?: any,
+) => boolean | null | undefined;